many changes

+++ src/backup.py
@@ -2,50 +2,55 @@
 
 import gzip
 import hashlib
-import helpers
 import logging
 import os
-import server
-import settings
 import shutil
 import time
 
-INCREMENTAL_PATH = os.path.join(settings.BACKUPS_PATH, "incremental")
+import helpers
+import server
+import settings
+import world
 
-# create backups-path
-if not os.path.exists(INCREMENTAL_PATH):
-    os.mkdir(INCREMENTAL_PATH)
-    os.chmod(INCREMENTAL_PATH, 0o777)
-
-def create(server_name, level_name=None, backup_name=None, description=None, compress=False):  #30xx
+def create(server_name=None, level_name=None, backup_name=None, overwrite=False, description=None, compress=False):  #201x
     if helpers.is_empty(server_name):
-        return 'server-name is required', 3001
-    
-    if server.is_running(server_name):
-        return 'this server is running', 3002
-    
-    result = server.get_world(server_name, level_name)
-    if result[1] > 0:
-        return 'cannot create backup', 3003, result
-    level_name = result[0]
+        return 'server-name is required', 2011
 
-    result = server.get_server_version(server_name)
+    server_path = os.path.join(settings.SERVER_PATH, server_name)
+    if not os.path.exists(server_path):
+        return f"server with this name '{server_name}' does not exists", 2012
+    
+    result = world.is_running(server_name, level_name)
     if result[1] > 0:
-        return 'cannot create backup', 3004, result
-    server_version = result[0]
+        return 'error by running-check', 2013, result
+    elif result[0]['state']:
+        return 'world is in use', 2014
+    level_name = result[0]['level-name']
 
     current_time = time.time()
     if helpers.is_empty(backup_name):
-        backup_name = time.strftime('%Y%m%d_%H%M%S', time.localtime(current_time))
-    backup_file = os.path.join(settings.BACKUPS_PATH, backup_name)
-    if os.path.exists(backup_file):
-        return 'cannot create backup', 3004, result
-    
-    file_list = f"server-name={server_name}\nserver-version={server_version}\nlevel-name={level_name}\ndatetime={time.strftime('%Y-%m-%d %H:%M', time.localtime(current_time))}\n"
-    if description:
-        file_list += f"description={description}\n"
+        backup_name = time.strftime('%Y%m%d_%H%M%S', time.localtime(current_time)) + '_' + server_name + '_' + level_name
+    else:
+        backup_name = backup_name[:-len('.properties')] if backup_name.endswith('.properties') else backup_name
 
-    world_path = os.path.join(settings.SERVER_PATH, server_name, 'worlds', level_name)
+    backup_file = os.path.join(settings.BACKUPS_PATH, backup_name + '.properties')
+    if os.path.exists(backup_file) and not helpers.is_true(overwrite):
+        return 'backup with this name already exists', 2015, result
+
+    backup_properties = {
+        'server-name': server_name,
+        'level-name': level_name,
+        'backup-name': backup_name,
+        'datetime': time.strftime('%Y-%m-%d %H:%M', time.localtime(current_time)),
+        'description': description
+    }
+    result_properties = backup_properties.copy()
+    result_properties['file-count'] = 0
+
+    result = server.get_version(server_name)
+    backup_properties['version'] = result[0]['version'] if result[1] == 0 else 'unknown'
+
+    world_path = os.path.join(server_path, 'worlds', level_name)
     try:
         for root, _, files in os.walk(world_path):
             for file in files:
@@ -53,8 +58,9 @@ def create(server_name, level_name=None, backup_name=None, description=None, com
                 relative_file_path = os.path.relpath(input_file, world_path)
                 with open(input_file, 'rb') as f:
                     file_hash = hashlib.sha512(f.read()).hexdigest()
-                file_list += f"{file_hash}={relative_file_path}\n"
-                output_file = os.path.join(INCREMENTAL_PATH, file_hash)
+                backup_properties[file_hash] = relative_file_path
+                result_properties['file-count'] += 1
+                output_file = os.path.join(settings.INCREMENTAL_PATH, file_hash)
                 if not os.path.exists(output_file):
                     if helpers.is_true(compress):
                         with open(input_file, 'rb') as f_in:
@@ -64,21 +70,27 @@ def create(server_name, level_name=None, backup_name=None, description=None, com
                         shutil.copy(input_file, output_file)
     except Exception as e:
         logging.error(f"unexpected error: {e}")
-        return str(e), 3005
+        return str(e), 2016
 
-    with open(backup_file, 'w') as f:
-        f.write(file_list)
+    # write backup
+    result = helpers.write_properties(backup_file, backup_properties)
+    if result[1] > 0:
+        return 'error at write backup', 2018, result
     
     helpers.change_permissions_recursive(settings.BACKUPS_PATH, 0o777)
-    return {'backup-name':backup_name},0
+    return result_properties, 0
 
-def restore(backup_name, server_name=None, level_name=None):  # 31xx
+def restore(backup_name=None, server_name=None, level_name=None):  # 202x
     if helpers.is_empty(backup_name):
-        return 'backup-name is required', 3101
+        return 'backup-name is required', 2021
     
-    result = helpers.read_properties(os.path.join(settings.BACKUPS_PATH, backup_name))
+    result = _get_backup_by_name(backup_name)
     if result[1] > 0:
-        return 'cannot read backup-details', 3102
+        return 'backup does not exists', 2022, result
+    
+    result = helpers.read_properties(os.path.join(settings.BACKUPS_PATH, result[0]['backup-file']))
+    if result[1] > 0:
+        return 'cannot read backup-details', 2023
     properties = result[0]
 
     if helpers.is_empty(server_name):
@@ -86,24 +98,26 @@ def restore(backup_name, server_name=None, level_name=None):  # 31xx
     
     server_path = os.path.join(settings.SERVER_PATH, server_name)
     if not os.path.exists(server_path):
-        return 'server-name not exists', 3103
-    
-    if server.is_running(server_name):
-        return 'this server is running', 3104
+        return f'server not exists: {server_name}', 2024
 
     if helpers.is_empty(level_name):
         level_name = properties['level-name']
-    
+
+    result = world.is_running(server_name, level_name)
+    if result[1] == 0 and result[0]['state']:
+        return 'world is in use', 2025
+
+    # begin restoring
     for sha512 in properties.items():
-        if len(sha512) == 128 and not os.path.exists(os.path.join(INCREMENTAL_PATH, sha512)):
-            return 'backup is corrupt. you should delete it.', 3105
+        if len(sha512) == 128 and not os.path.exists(os.path.join(settings.INCREMENTAL_PATH, sha512)):
+            return 'backup is corrupt. you should delete it.', 2026
 
     world_path = os.path.join(server_path, 'worlds', level_name)
     if os.path.exists(world_path):
         # remove old existing world
         result = helpers.remove_dirtree(world_path)
         if result[1] > 0:
-            return 'cannot remove world', 3106, result
+            return 'cannot remove world', 2027, result
 
     def is_gzip_file(file_path):
         try:
@@ -120,7 +134,7 @@ def restore(backup_name, server_name=None, level_name=None):  # 31xx
                 if not os.path.exists(os.path.dirname(output_file)):
                     os.makedirs(os.path.dirname(output_file))
 
-                input_file = os.path.join(INCREMENTAL_PATH, sha512)
+                input_file = os.path.join(settings.INCREMENTAL_PATH, sha512)
                 if is_gzip_file(input_file):
                     with gzip.open(input_file, 'rb') as f_in:
                         with open(output_file, 'wb') as f_out:
@@ -133,36 +147,46 @@ def restore(backup_name, server_name=None, level_name=None):  # 31xx
                         level_file.write(level_name)
     except Exception as e:
         logging.error(f"unexpected error: {e}")
-        return str(e), 3107
+        return str(e), 2028
     
     helpers.change_permissions_recursive(world_path, 0o777)
-    return 'restored', 0
+    return {
+        'server-name': server_name,
+        'backup-name': backup_name,
+        'level-name': level_name,
+        'state': 'restored'
+    }, 0
 
-def list():  #32xx
+def list():  #203x
     result_list = {}
     for backup_file in [file for file in os.listdir(settings.BACKUPS_PATH) if os.path.isfile(os.path.join(settings.BACKUPS_PATH, file))]:
         result = helpers.read_properties(os.path.join(settings.BACKUPS_PATH, backup_file))
         if result[1] > 0:
-            return 'cannot read backup-details', 3201
-        properties = result[0]
-        result_list[backup_file] = {
-            'backup-name': backup_file,
-            'server-name': properties['server-name'],
-            'server-version': properties['server-version'],
-            'level-name': properties['level-name'],
-            'datetime': properties['datetime'],
-            'description': properties['description'] if 'description' in properties else None
-        }
+            return 'cannot read backup-details', 2031
+        
+        properties = {'file-count': 0}
+        for key, value in result[0].items():
+            if len(key) == 128:
+                properties['file-count'] += 1
+            elif key != 'file-count':
+                properties[key] = value
+
+        backup_file = backup_file[:-len('.properties')] if backup_file.endswith('.properties') else backup_file
+        result_list[backup_file] = properties
     return result_list, 0
     
-def remove(backup_name):  # 33xx
+def remove(backup_name=None):  # 204x
     if helpers.is_empty(backup_name):
-        return 'backup-name is required', 3301
+        return 'backup-name is required', 2041
     
+    result = _get_backup_by_name(backup_name)
+    if result[1] > 0:
+        return 'backup does not exists', 2042, result
+
     # remove backup-file
-    backup_file = os.path.join(settings.BACKUPS_PATH, backup_name)
+    backup_file = os.path.join(settings.BACKUPS_PATH, result[0]['backup-file'])
     if not os.path.exists(backup_file):
-        return f'backup does not exists {backup_name}', 3303
+        return f'backup does not exists {backup_name}', 2043
     os.remove(backup_file)
     
     # list all referenced sha512-files
@@ -170,7 +194,7 @@ def remove(backup_name):  # 33xx
     for backup_file in [file for file in os.listdir(settings.BACKUPS_PATH) if os.path.isfile(os.path.join(settings.BACKUPS_PATH, file))]:
         result = helpers.read_properties(os.path.join(settings.BACKUPS_PATH, backup_file))
         if result[1] > 0:
-            return 'cannot read backup-details', 3304
+            return 'cannot read backup-details', 2044
         properties = result[0]
         for sha512 in properties:
             if len(sha512) == 128 and sha512 not in linked_files:
@@ -178,10 +202,37 @@ def remove(backup_name):  # 33xx
 
     # remove all unreferenced sha512-files
     try:
-        for sha512 in os.listdir(INCREMENTAL_PATH):
+        for sha512 in os.listdir(settings.INCREMENTAL_PATH):
             if sha512 not in linked_files:
-                os.remove(os.path.join(INCREMENTAL_PATH, sha512))
-        return 'removed', 0
+                os.remove(os.path.join(settings.INCREMENTAL_PATH, sha512))
+        return {
+            'backup-name': result[0]['backup-name'],
+            'backup-file': result[0]['backup-file'],
+            'state': 'removed'
+        }, 0
     except Exception as e:
         logging.error(f"unexpected error: {e}")
-        return str(e), 3305
+        return str(e), 2045
+
+def _get_backup_by_name(backup_name):  # 205x
+    backup_name = backup_name[:-len('.properties')] if backup_name.endswith('.properties') else backup_name
+    backup_file = backup_name + '.properties'
+    
+    if os.path.exists(os.path.join(settings.BACKUPS_PATH, backup_file)):
+        return {
+            'backup-name': backup_name,
+            'backup-file': backup_file
+        }, 0
+
+    result = list()
+    if result[1] > 0:
+        return 'cannot find backup', 2051, result
+
+    for file, properties in result[0].items():
+        if file == backup_name or file == backup_file or properties['backup-name'] == backup_name or properties['backup-name'] == backup_file:
+            return {
+                'backup-name': properties['backup-name'],
+                'backup-file': file
+            }, 0
+
+    return 'cannot found backup', 2052
